Refactor and add proper CLI
Uses optparse-applicative

Also drops requirement for `cabal.cfg`. As we have the knowledge about the content of `cabal.cfg` anyhow. We can just compute it in a temporary directory and use it, instead of requiring the
user to provide redundant information, which could lead to inconsistencies and
hard to find issues / bad experience.

Fixes #3.